Fixed moveToAngle logic so it works for every min/max angle and made the makeShotAtCloseRangeAndMoveTurretToHome not call setStateHoming but instead use MoveToAngle(0)

+++ src/main/java/robot/turret/BasicTurretComponentsA.java
@@ -61,7 +61,7 @@ public class BasicTurretComponentsA implements TurretComponents {
   }
 
   @Override
-  public double getAbsoluteEncoderOffset() {
+  public int getAbsoluteEncoderOffset() {
     return ABSOLUTE_ENCODER_OFFSET;
   }
 }
